chore: move augment_dst to be able to share as utils
